First Visual Studio C# Application
Requirements
Console Application New Project Window Example
Create a new Visual C# Console App using .NET Framework project and name it ConsoleApplication.
Modify the default code to match the first code window in The Hello World Project in lesson one of the tutorial except for the name of the namespace.
Add two more WriteLine statements directly after the Console.WriteLine("Hello World!"); statement that display 1) your full name, and 2) your location in terms of state or country.
Add a line of code that reproduces the Console.WriteLine($"Hello {name}!"); using standard concatenation where name is a variable containing your first name. Hint: Use +
Add a line the writes the current date.
Add a line that displays and labels the number of days until Christmas.
Add the program example from section 2.1 of the C# Programming Yellow Book in the required readings. In addition to the given code, include the following additions: 
provide appropriate text when requesting dimensional input, and
cause the program to pause in the console so that the application does not automatically terminate when launched from the Visual Studio run debugger tool.